+++ docs/changelog.md
@@ -140,6 +140,8 @@ Many feature gates are removed in order to make most features make always availa
 - define `_std_core` separately and privately per crate.
 - update `CONST!` macro with new arms: `hidden macro_export`, `inline macro_export`.
 - add `location` arms to the `_doc!` macro.
+- update `impl_traits!` macro:
+  - add new arms for: `PartialEq`.
 - update `const_assert!` macro:
   - add new arms: `ne_buf`, `ne_str`.
   - add support for comparing slices of primitives and slices of slices of primitives.